requirements.txt
numpy>=1.21.0
pandas>=1.3.0
pyodbc>=4.0.32
pyyaml>=6.0
requests>=2.26.0
azure-quantum>=0.27.0
azure-storage-blob>=12.9.0
python-dotenv>=0.19.0
pytest>=7.0.0
black>=22.3.0
flake8>=4.0.0 